ex/pointer-revocation: mark pointers volatile
Some combination of "assert(obj1 == obj2)" and the UBs of using a pointer value
after free and the overtly visible UAF was giving clang license to conflate obj1
and obj2 post-assert, resulting in SIGPROT being delivered too early, at least,
as far as the source appeared to be concerned (but again, UB).  Marking the
pointers (not the pointees) volatile turns off enough such reasoning that the
trap now again happens at the source-visible reuse of obj1.