Introduce do_cat().
do_cat() is an example of a single interface that calls a type-specific
interface depending on an object type. The object type is determined by
a separate argument and the object can be referenced with a pointer or
an integer value.

do_cat() accepts a file descriptor or a FILE pointer and uses
verbose_cat() or raw_cat() depending on a value of a separate boolean
argument.

This change deliberately introduces a CHERI capability misuse bug that
is to be presented in the adapt-c exercise.