Introduce write_pos().
write_pos() behaves as write(2) buf splits the buf argument into two
arguments: an address of a source buffer and an offset within the
buffer. write_pos() copies nbyte bytes starting from the offset in
the buffer.

This change deliberately introduces a CHERI ambiguous provenance bug
that is to be presented in the adapt-c exercise.